requirements.txt
notebook==6.5.4
numpy~=1.24.4
pandas~=2.0.3
scikit-learn~=1.3.2
transformers~=4.33.3
huggingface_hub
nltk~=3.8.1
checklist~=0.0.11
spacy~=3.7.2
gensim~=4.3.2
sklearn==0.0.post9
pandarallel==1.6.5
chardet
baidu-aip
torch~=1.13.1+cu116
networkx~=3.1
bert_score
stanfordnlp==0.2.0
tqdm==4.66.1
simpletransformers==0.64.5
levenshtein==0.23.0
keybert==0.8.3
boto3==1.34.11
azure-ai-textanalytics==5.2.0